requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.1
arviz==0.21.0
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
boto3==1.37.25
botocore==1.37.25
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cohere==5.14.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.3.9
distro==1.9.0
Django==5.2
django-import-export==4.3.7
dm-tree==0.1.9
eval_type_backport==0.2.2
faiss-cpu==1.10.0
fake-useragent==2.1.0
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastavro==1.10.0
fastprogress==1.0.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.7.2
groq==0.21.0
grpcio==1.71.0
grpcio-status==1.71.0
gymnasium==1.1.1
h11==0.14.0
h5netcdf==1.6.1
h5py==3.13.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.22
lightgbm==4.6.0
llvmlite==0.44.0
logfire-api==3.12.0
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
mpmath==1.3.0
multidict==6.3.1
mysqlclient==2.2.7
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai==1.70.0
opentelemetry-api==1.31.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parsel==1.10.0
patsy==1.0.1
pillow
prompt_toolkit==3.0.50
propcache==0.3.1
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-ai==0.0.49
pydantic-ai-slim==0.0.49
pydantic-evals==0.0.49
pydantic-graph==0.0.49
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyDispatcher==2.0.7
Pygments==2.19.1
pymc3==3.11.4
PyMySQL==1.1.1
pyod==2.0.4
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
semver==3.0.4
sentence-transformers==4.0.2
service-identity==24.2.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
sse-starlette==2.2.1
stable_baselines3==2.6.0
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
tablib==3.8.0
tenacity==9.0.0
tensorflow-probability==0.25.0
Theano-PyMC==1.1.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
torch==2.6.0
torch-geometric==2.6.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
Twisted==24.11.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xarray==2025.3.1
xarray-einstats==0.8.0
xgboost==3.0.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
